🐛 fix(monitor): correct static_dir default and format specifier
- static_dir default was "www", changed to "static" to match directory structure
- capture response used %zu (unsupported by mg_http_reply), changed to %lu
- added .gitignore and Makefile help target

+++ tools/crt_monitor/routes.c
@@ -73,8 +73,8 @@ static void handle_api_capture(struct mg_connection *c,
     fclose(f);
 
     mg_http_reply(c, 200, "Content-Type: application/json\r\n",
-                  "{\"ok\":true,\"file\":\"%s\",\"bytes\":%zu}\n",
-                  filename, len);
+                  "{\"ok\":true,\"file\":\"%s\",\"bytes\":%lu}\n",
+                  filename, (unsigned long)len);
 }
 
 /* ------------------------------------------------------------------ */
